Add $libtool_no_delete_archives, which doesn't remove .la files.
--HG--
extra : convert_revision : 6f3e927947d6d3c7641bad572e7c021bee2f4f32

+++ shutils/install_funcs.sh
@@ -74,7 +74,9 @@ install_src_phase()
 	run_func post_install
 
 	# Remove libtool archives from pkg destdir.
-	find ${DESTDIR} -type f -name \*.la -delete
+	if [ -n "$libtool_no_delete_archives" ]; then
+		find ${DESTDIR} -type f -name \*.la -delete
+	fi
 
 	# unset cross compiler vars.
 	[ -n "$cross_compiler" ] && cross_compile_unsetvars
